Add a function for bulk PMU event discovery
Currently, the only way to see if an event is supported is to attempt to
configure it on some counter. This requires a separate SBI call for each
event, and requires all counters to be otherwise free. Add a new
function to allow checking event support in bulk (XLEN events at a
time), regardless of which counters are currently in use.

Signed-off-by: Samuel Holland <samuel.holland@sifive.com>