Require misaligned access handling by default if implemented
Some SBI implementations may not be able to emulate misaligned accesses,
and therefore may always delegate the exception to supervisor software.
However, if the implementation does allow supervisor software to disable
this delegation, then it must do so by default for compatibility with
existing supervisor software and the UEFI specification.

Suggested-by: Heinrich Schuchardt <heinrich.schuchardt@canonical.com>
Signed-off-by: Samuel Holland <samuel.holland@sifive.com>